pyproject.toml
[tool.poetry]
name = "drf-cci-rebalancer"
version = "0.1.0"
description = ""
authors = ["NinaNikitenko <ninanineliya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
django = "4.2.9"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.8"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
coverage = "^7.6.1"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
